combine validation and fixed
