+++ src/main/resources/static/js/register.js
@@ -1,9 +1,16 @@
 document.addEventListener("DOMContentLoaded", () => {
     const registerForm = document.getElementById("registerForm");
+    const ownerRegisterForm = document.getElementById("ownerRegisterForm");
 
-    if (registerForm) {
-        registerForm.addEventListener("submit", (e) => {
-            e.preventDefault();
+    const form = registerForm || ownerRegisterForm;
+    if(!form) return;
+
+    const currentPathType = window.location.pathname.includes('owner') ? 'owner' : 'member';
+    let isBusinessNumberValid = false;
+
+
+    form.addEventListener("submit", (e) => {
+        e.preventDefault();
 
             // 모든 검증
             checkName(document.getElementById("registerName"));
@@ -13,10 +20,19 @@ document.addEventListener("DOMContentLoaded", () => {
             checkPwd(document.getElementById("registerPwd"));
             checkPwdConfirm();
 
-            setTimeout(() => {
-                const invalidInputs = registerForm.querySelectorAll(".is-invalid");
+            if (currentPathType === 'owner') {
+                checkBusinessNum(document.getElementById("businessNum"));
+            }
+
+
+        setTimeout(() => {
+                const invalidInputs = form.querySelectorAll(".is-invalid");
                 if (invalidInputs.length > 0) {
                     alert("입력 항목을 다시 확인해주세요.");
+                    return;
+                }
+                    if(currentPathType === 'owner' && !isBusinessNumberValid){
+                    alert("사업자 번호 조회를 진행해주세요.");
                     return;
                 }
 
@@ -28,7 +44,12 @@ document.addEventListener("DOMContentLoaded", () => {
                     password: document.getElementById("registerPwd").value,
                     passwordConfirm: document.getElementById("registerPwdConfirm").value
                 };
-                fetch("/api/member/register", {
+
+                    if(currentPathType === 'owner') {
+                        data.businessNum = document.getElementById("businessNum").value;
+                        data.dinerName =  document.getElementById("ownerDinerName").value;
+                    }
+                fetch(`/api/${currentPathType}/register`, {
                     method: "POST",
                     headers: { "Content-Type": "application/json" },
                     body: JSON.stringify(data)
@@ -45,5 +66,54 @@ document.addEventListener("DOMContentLoaded", () => {
                     .catch(err => console.error("Register Error:", err));
             }, 600);
         });
+
+//사업자 번호로 조회
+const businessNumberBtn = document.getElementById("businessNumBtn");
+if (businessNumberBtn) {
+    businessNumberBtn.addEventListener("click", () => {
+        const businessNumber = document.getElementById("businessNum").value.trim();
+        const reg = /^\d{10}$/;
+        if (businessNumber === "") {
+            alert("사업자 번호를 입력해주세요.");
+            return;
+        } else if(businessNumber.length !== 10 || !reg.test(businessNumber)){
+            alert("올바른 사업자 번호 형식이 아닙니다. 숫자 10자리로 입력해주세요.");
+            return;
+        } else {
+            fetchBusinessInfo(businessNumber);
+        }
+    });
+}
+
+async function fetchBusinessInfo(businessNumber) {
+
+    try{
+        // Controller의 경로(@RequestMapping + @GetMapping)에 맞춰 수정
+        const url = `/api/owner/proxy/business-info?query=${businessNumber}`;
+        const res = await fetch(url);
+        if(res.ok){
+            const data = await res.json();
+            console.log(data);
+            const items = data.items || data.data;
+            if(items && items.length > 0){
+                const info = items[0];
+
+                // 받아온 상호명을 공백 제거
+                const cleanDinerName = info.company.replace(/\s+/g, '');
+                document.getElementById("ownerDinerName").value = cleanDinerName;
+                isBusinessNumberValid = true;
+                document.getElementById("businessNum").readOnly = true;
+                alert("사업자 정보가 정상적으로 조회되었습니다.");
+            } else {
+                alert("유효한 사업자 번호가 아닙니다.");
+            }
+        } else{
+            const error = await res.text();
+            alert("사업자 정보 조회에 실패했습니다." + error);
+        }
+    } catch(err){
+        console.error("사업자 정보 조회 오류:", err);
+        alert("사업자 정보 조회 중 오류가 발생했습니다.");
     }
+}
 });
