+++ src/main/resources/templates/user/owner-register.html
@@ -23,8 +23,8 @@
         <form id="ownerRegisterForm" method="post">
 
           <div class="form-floating mb-3">
-            <input type="text" class="form-control" id="ownerRegisterName" placeholder="이름" onkeyup="checkName(this)">
-            <label for="ownerRegisterName">이름 <span class="text-danger">*</span></label>
+            <input type="text" class="form-control" id="registerName" placeholder="이름" onkeyup="checkName(this)">
+            <label for="registerName">이름 <span class="text-danger">*</span></label>
             <div id="error-registerName" class="small ms-1"></div>
           </div>
 
@@ -37,8 +37,8 @@
           </div>
 
           <div class="form-floating mb-3">
-            <input type="text" class="form-control" id="ownerRegisterId" placeholder="아이디" onkeyup="checkUsername(this)">
-            <label for="ownerRegisterId">아이디 <span class="text-danger">*</span></label>
+            <input type="text" class="form-control" id="registerId" placeholder="아이디" onkeyup="checkUsername(this, 'owner')">
+            <label for="registerId">아이디 <span class="text-danger">*</span></label>
             <div id="error-registerId" class="small ms-1"></div>
           </div>
 
@@ -46,10 +46,10 @@
             <label class="form-label ms-1">이메일 <span class="text-danger">*</span></label>
             <input type="hidden" id="totalEmail">
             <div class="input-group">
-              <input type="text" class="form-control" id="ownerRegisterEmail" placeholder="이메일" onkeyup="EmailValidation()">
+              <input type="text" class="form-control" id="registerEmail" placeholder="이메일" onkeyup="EmailValidation('owner')">
               <span class="input-group-text">@</span>
-              <input type="text" class="form-control" id="ownerRegisterDomain" placeholder="도메인" onkeyup="EmailValidation()">
-              <select class="form-select" id="ownerEmailDomainSelect" onchange="selectDomain()">
+              <input type="text" class="form-control" id="registerDomain" placeholder="도메인" onkeyup="EmailValidation('owner')">
+              <select class="form-select" id="emailDomainSelect" onchange="selectDomain()">
                 <option value="">직접 입력</option>
                 <option value="naver.com">naver.com</option>
                 <option value="gmail.com">gmail.com</option>
@@ -61,21 +61,21 @@
           </div>
 
           <div class="mb-3">
-            <label for="ownerRegisterPhone" class="form-label ms-1">전화번호(선택)</label>
-            <input type="text" class="form-control" id="ownerRegisterPhone" name="phone"
+            <label for="registerPhone" class="form-label ms-1">전화번호(선택)</label>
+            <input type="text" class="form-control" id="registerPhone" name="phone"
                    placeholder="( - )제외, 번호만 입력해주세요." onkeyup="checkPhone(this)">
             <div id="error-registerPhone" class="small ms-1 mt-1"></div>
           </div>
 
           <div class="form-floating mb-3">
-            <input type="password" class="form-control" id="ownerRegisterPwd" onkeyup="checkPwd(this)">
-            <label for="ownerRegisterPwd">비밀번호 <span class="text-danger">*</span></label>
+            <input type="password" class="form-control" id="registerPwd" onkeyup="checkPwd(this)">
+            <label for="registerPwd">비밀번호 <span class="text-danger">*</span></label>
             <div id="error-registerPwd" class="small ms-1"></div>
           </div>
 
           <div class="form-floating mb-4">
-            <input type="password" class="form-control" id="ownerRegisterPwdConfirm" onkeyup="checkPwdConfirm()">
-            <label for="ownerRegisterPwdConfirm">비밀번호 확인 <span class="text-danger">*</span></label>
+            <input type="password" class="form-control" id="registerPwdConfirm" onkeyup="checkPwdConfirm()">
+            <label for="registerPwdConfirm">비밀번호 확인 <span class="text-danger">*</span></label>
             <div id="error-registerPwdConfirm" class="small ms-1"></div>
           </div>
           <button type="submit" class="btn btn-primary w-100 py-3 fw-bold">가입 완료</button>
@@ -88,7 +88,7 @@
   <div th:insert="~{fragment/common :: footer}"></div>
 </footer>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
-<script th:src="@{/js/ownerValidation.js}" type="text/javascript"></script>
-<script th:src="@{/js/ownerRegister.js}"></script>
+<script src="/js/validation.js" type="text/javascript"></script>
+<script th:src="@{/js/register.js}"></script>
 </body>
 </html>
