+++ src/main/resources/templates/user/ownerPage.html
@@ -289,6 +289,6 @@
 </footer>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js" crossorigin="anonymous"></script>
 <script src = "/js/ownerPage.js" type="text/javascript"></script>
-<script src = "/js/ownerValidation.js" type="text/javascript"></script>
+<script src="/js/validation.js" type="text/javascript"></script>
 </body>
 </html>
